Setup leptos template
