+++ Cargo.toml
@@ -4,10 +4,13 @@ version = "0.1.0"
 authors = ["sorz <orz@sorz.org>"]
 edition = "2024"
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
+leptos = { version = "0.7", features = ["csr", "nightly"] }
+leptos_meta = { version = "0.7" }
+leptos_router = { version = "0.7", features = ["nightly"] }
+console_log = "1"
+log = "0.4"
+console_error_panic_hook = "0.1"
 wasm-bindgen = "0.2"
 web-sys = { version = "0.3", features = ["console", "Blob", "BlobPropertyBag"] }
 js-sys = { version = "0.3" }
