+++ rust-toolchain.toml
@@ -1,0 +1,3 @@
+
+[toolchain]
+channel = "nightly"
