+++ src/app/pages/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod home;
+pub(crate) mod not_found;
