+++ src/app/pages/not_found.rs
@@ -1,0 +1,6 @@
+use leptos::prelude::*;
+
+#[component]
+pub fn NotFound() -> impl IntoView {
+    view! { <h1>"Uh oh!" <br /> "We couldn't find that page!"</h1> }
+}
