+++ src/bin/worker.rs
@@ -1,0 +1,7 @@
+fn main() {
+    console_error_panic_hook::set_once();
+    console_log::init_with_level(log::Level::Debug).unwrap();
+    log::debug!("staring service worker");
+
+    // TODO
+}
